Modification du type de la valeur de retour de effectuerCoup de l'IA en raison de problémes sur machines 64 bits.

+++ include/ia.h
@@ -25,7 +25,7 @@ public:
         //this->nom = nom;
     }
 
-    unsigned int effectuerCoup() {
+    size_t effectuerCoup() {
         return effectuerCoup(profondeur);
     }
